Fix test, move bib info services
