+++ src/main/java/org/folio/innreach/domain/service/impl/BibInfoServiceImpl.java
@@ -1,4 +1,4 @@
-package org.folio.innreach.external.service.impl;
+package org.folio.innreach.domain.service.impl;
 
 import static java.util.List.of;
 
@@ -14,7 +14,7 @@ import org.folio.innreach.domain.service.InventoryViewService;
 import org.folio.innreach.dto.BibInfoResponseDTO;
 import org.folio.innreach.external.dto.BibInfoResponse;
 import org.folio.innreach.external.mapper.InnReachResponseMapper;
-import org.folio.innreach.external.service.BibInfoService;
+import org.folio.innreach.domain.service.BibInfoService;
 
 @Log4j2
 @RequiredArgsConstructor
