+++ src/test/java/org/folio/innreach/external/service/impl/BibInfoServiceImplTest.java
@@ -20,6 +20,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.folio.innreach.domain.service.CentralServerService;
 import org.folio.innreach.domain.service.InstanceTransformationService;
 import org.folio.innreach.domain.service.InventoryViewService;
+import org.folio.innreach.domain.service.impl.BibInfoServiceImpl;
 import org.folio.innreach.dto.BibInfo;
 import org.folio.innreach.dto.CentralServerDTO;
 import org.folio.innreach.dto.Instance;
